Get rid of cer files, use new shared helper
